Update BlockLocator.py
Fixed bug in BlockLocator 

(used by `findFunctionByContainedAddress` and `findBlockByContainedAddress`)

It was assumed the lowerbound check was done by the `bisect` call. This is true for almost all blocks, however,  a special case must be handled for the first block. 

When an address smaller than the first block was queried, the BlockLocator would erroneously wrap around and return the last block in the program.